feat(FilesTableFilters): separate filters from the table reload
