+++ src/sections/dataset/dataset-files/DatasetFiles.tsx
@@ -1,11 +1,10 @@
-import { useFilesTable } from './files-table/useFilesTable'
 import { FileRepository } from '../../../files/domain/repositories/FileRepository'
-import { useFiles } from './useFiles'
-import { useEffect, useState } from 'react'
+import { useState } from 'react'
 import { FilesTable } from './files-table/FilesTable'
-import { SpinnerSymbol } from './files-table/spinner-symbol/SpinnerSymbol'
 import { FileCriteriaControls } from './file-criteria-controls/FileCriteriaControls'
-import { FileCriteria } from '../../../files/domain/models/FileCriteria'
+import { FileAccessOption, FileCriteria, FileTag } from '../../../files/domain/models/FileCriteria'
+import { FilesCountInfo } from '../../../files/domain/models/FilesCountInfo'
+import { FileType } from '../../../files/domain/models/File'
 
 interface DatasetFilesProps {
   filesRepository: FileRepository
@@ -14,6 +13,27 @@ interface DatasetFilesProps {
 }
 
 const MINIMUM_FILES_TO_SHOW_CRITERIA_INPUTS = 2
+const filesCountInfo: FilesCountInfo = {
+  total: 222,
+  perFileType: [
+    {
+      type: new FileType('text'),
+      count: 5
+    },
+    {
+      type: new FileType('image'),
+      count: 485
+    }
+  ],
+  perAccess: [
+    { access: FileAccessOption.PUBLIC, count: 222 },
+    { access: FileAccessOption.RESTRICTED, count: 10 }
+  ],
+  perFileTag: [
+    { tag: new FileTag('document'), count: 5 },
+    { tag: new FileTag('code'), count: 10 }
+  ]
+} // TODO: Get from use case
 
 export function DatasetFiles({
   filesRepository,
@@ -21,24 +41,8 @@ export function DatasetFiles({
   datasetVersion
 }: DatasetFilesProps) {
   const [criteria, setCriteria] = useState<FileCriteria>(new FileCriteria())
-  const { files, filesCountInfo, isLoading } = useFiles(
-    filesRepository,
-    datasetPersistentId,
-    datasetVersion,
-    criteria
-  )
-  const { table, setFilesTableData } = useFilesTable()
-
   const handleCriteriaChange = (newCriteria: FileCriteria) => {
     setCriteria(newCriteria)
-  }
-
-  useEffect(() => {
-    setFilesTableData(files)
-  }, [files])
-
-  if (isLoading) {
-    return <SpinnerSymbol />
   }
 
   return (
@@ -50,7 +54,12 @@ export function DatasetFiles({
           filesCountInfo={filesCountInfo}
         />
       )}
-      <FilesTable table={table} />
+      <FilesTable
+        filesRepository={filesRepository}
+        datasetPersistentId={datasetPersistentId}
+        datasetVersion={datasetVersion}
+        criteria={criteria}
+      />
     </>
   )
 }
